Move constants to cli
