+++ src/cli/commands/version.js
@@ -1,5 +1,3 @@
-'use strict';
-
 import chalk from 'chalk';
 import fs from 'fs';
