+++ src/cli/main.js
@@ -1,7 +1,5 @@
-'use strict';
-
 import chalk from 'chalk';
-import {Cli} from './Cli.js';
+import {commandMap} from './commandMap.js';
 
 const args = process.argv.slice(2);
 const command = args[0];
@@ -9,14 +7,14 @@ const params = args.slice(1);
 
 if (!command) {
   console.log(chalk.green('No command passed'));
-  Cli['--help'].run();
+  commandMap['--help'].run();
   process.exit(1);
 }
 
-if (!Cli[command]) {
+if (!commandMap[command]) {
   console.log(chalk.green(`Unknown command: ${command}`));
-  Cli['--help'].run();
+  commandMap['--help'].run();
   process.exit(1);
 }
 
-Cli[command].run(params);
+commandMap[command].run(params);
